Factor FS, Compiler logic out
The separation is not perfect, but at least now we can see a bit more of
what each part is doing. There are a lot of FIXMEs, TODOs and NOTEs to
remind us of what the assumptions are, but the code should really just
read more easily (which is hard when using Clang's APIs).
