--- src/ffi/compiler.cc
@@ -1,7 +1,0 @@
-// Copyright Microsoft and Project Verona Contributors.
-// SPDX-License-Identifier: MIT
-
-#include "compiler.h"
-
-namespace verona::ffi::compiler
-{} // namespace verona::ffi::compiler
